+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [1.1.37](https://github.com/jensvogt/euclid/compare/1.1.36...v1.1.37) (2026-09-24)
+
+
+### Bug Fixes
+
+* add bucket priority ([40865cb](https://github.com/jensvogt/euclid/commit/40865cb2596cf8cf96f2de55f69662f9266051d9))
+* add cpp example ([8064eb9](https://github.com/jensvogt/euclid/commit/8064eb9b100e74fdca3c9f862ebef64043c12833))
+* add cpp example ([9f8a2b1](https://github.com/jensvogt/euclid/commit/9f8a2b1ebbcc075e71f837f1392a77a0ca1e824d))
+
+## [1.1.36](https://github.com/jensvogt/euclid/compare/1.1.35...v1.1.36) (2026-09-23)
+
+
+### Bug Fixes
+
+* exists queries ([e73f7f0](https://github.com/jensvogt/euclid/commit/e73f7f00df13ea082616f607d6c9fa2e3342bda5))
+
+## [1.1.35](https://github.com/jensvogt/euclid/compare/1.1.34...v1.1.35) (2026-09-23)
+
+
+### Bug Fixes
+
+* graceful shutdown ([d2927b4](https://github.com/jensvogt/euclid/commit/d2927b494b63d543ca3e2e191eaa6a7ba04d0ab7))
+
 ## [1.1.34](https://github.com/jensvogt/euclid/compare/1.1.33...v1.1.34) (2026-09-22)